+++ v2/src/device/driver.h
@@ -86,7 +86,7 @@ typedef struct  xkblas_driver_t
     ////////////////////////////////
 
     /* alllocate and initialize a stream */
-    xkblas_stream_t * (*f_stream_create)(xkblas_stream_type_t type, unsigned int capacity);
+    xkblas_stream_t * (*f_stream_create)(xkblas_stream_type_t type, uint16_t capacity);
 
     /* deallocate a stream */
     void (*f_stream_delete)(xkblas_stream_t * istream);
